+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWBPFPBVJXG3NEWTHV482KRV"
+        "01KWBSJEV1DC6JKP3E91M70JNS"
     )
     r"""The unique identifier of the data source"""