+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -199,9 +199,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-30T07:21:18.786Z"
+    created: Optional[str] = "2026-06-30T08:15:17.240Z"
 
-    updated: Optional[str] = "2026-06-30T07:21:18.786Z"
+    updated: Optional[str] = "2026-06-30T08:15:17.240Z"
 
     updated_by_id: OptionalNullable[str] = UNSET