+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-30T07:21:16.506Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-30T08:15:15.305Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")