+++ packages/orq-rc/RELEASES.md
@@ -9099,3 +9099,13 @@ Based on:
 - [python v4.11.0-rc.59] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.59] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.59 - packages/orq-rc
+
+## 2026-06-30 08:18:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.60] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.60] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.60 - packages/orq-rc